+++ lib/layouts.dart
@@ -15,5 +15,6 @@
 library;
 
 export 'src/enums.dart';
+export 'src/layouts/shuffle.dart';
 export 'src/models.dart';
 export 'src/layouts/layouts.dart';